// src/types/interfaces.ts
export interface ExcelReport {
    date: string
    plate_no: string
    model?: string
    w_nw: string
    repair_type: string
    bay_number: string
    km_check_up: string
    series: string
    as_c: string
    as_details: string
    aj_c: string
    aj_details: string
    customer_arrival: string
    appointment_time: string
    cat_guard: string
    cat_receptionist: string
    ticketing_time: string
    customer_wt: string
    sa: string
    sa_rt_s: string
    sa_rt_e: string
    reception_lead_time: string
    reception_idle_time: string
    technician_cit: string
    technician_idle_time_0: string
    technician_idle_time_1: string
    technician_idle_time_2: string
    repair_s: string
    repair_e: string
    repair_lead_time: string
    repair_idle_time: string
    carwash_s: string
    carwash_e: string
    carwash_lead_time: string
    release_date: string
    car_out_time: string
    total_em_lt: string
    total_lt: string
    promised_time: string
    otd: string
    em60_ach: string
    remarks: string
}